Select across markdown lines and scroll wide tables
- Each prose line rendered as its own Text, so a selection could
  never span lines; consecutive lines now merge into one block
- Expanding a review bot's per-file details block let its wide
  table demand full width and squeeze the panes around it; tables
  now scroll horizontally within the pane instead
- The parsing moves to a cross-file extension for type length
